Cast `TILE_SIZE` to `double` for consistent division in `COLLISION_DISTANCE_PVG`.

+++ pacman/src/main/java/com/example/model/Constants.java
@@ -30,7 +30,7 @@ public class Constants {
     public static final double CENTER_EPS_PX = 1.5;
 
     public final static int COUNTDOWN_DURATION_TICKS = 60;
-    public static final double COLLISION_DISTANCE_PVG = TILE_SIZE/2;
+    public static final double COLLISION_DISTANCE_PVG = (double) TILE_SIZE /2;
     public static final double COLLISION_DISTANCE_PVP = TILE_SIZE;
 
     public static final double FRUIT_RESPAWN_DELAY_SEC = 5.0;
